requirements.txt
meerk40t
pyserial